Create a WebAPI base class, and use it to build GitHubAPI and BitbucketAPI objects. Use these from GitHubProvider and BitbucketProvider
