+++ src/ServiceProviders/ProviderEnvironment.php
@@ -4,15 +4,73 @@ namespace Pantheon\TerminusBuildTools\ServiceProviders;
 
 /**
  * Store variables relevant to a provider.
+ *
+ * Use like an ArrayObject to cache transient variables.
+ *
+ * Use setToken / token to get / fetch variables that should be
+ * persisted with the CI service.
  */
-class ProviderEnvironment extends \ArrayObject
+class ProviderEnvironment extends \ArrayObject implements ServiceTokenStorage
 {
-
     const USER_AGENT = 'pantheon/terminus-build-tools-plugin';
+
+    protected $token_key = 'TOKEN';
+    protected $tokens = [];
+    protected $serviceName;
 
     public function __construct(array $data = [])
     {
         parent::__construct($data);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function hasToken($key = false)
+    {
+        if (!$key) {
+            $key = $this->token_key;
+        }
+        return in_array($key, $this->tokens);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function token($key = false)
+    {
+        if (!$key) {
+            $key = $this->token_key;
+        }
+        return $this[$key];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setToken($key, $token)
+    {
+        $this->token_key = $key;
+        $this->tokens[] = $key;
+        $this[$key] = $token;
+        return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function serviceName()
+    {
+        return $this->serviceName;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function setServiceName($serviceName)
+    {
+        $this->serviceName = $serviceName;
+        return $this;
     }
 
     /**
