+++ src/ServiceProviders/RepositoryProviders/Bitbucket/BitbucketProvider.php
@@ -7,11 +7,11 @@ use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 
 use Pantheon\TerminusBuildTools\Credentials\CredentialClientInterface;
-use Pantheon\TerminusBuildTools\Credentials\CredentialProviderInterface;
 use Pantheon\TerminusBuildTools\ServiceProviders\RepositoryProviders\GitProvider;
-use Pantheon\TerminusBuildTools\Credentials\CredentialRequest;
 use Pantheon\TerminusBuildTools\Utility\ExecWithRedactionTrait;
 use Pantheon\TerminusBuildTools\ServiceProviders\RepositoryProviders\RepositoryEnvironment;
+use Pantheon\TerminusBuildTools\API\Bitbucket\BitbucketAPI;
+use Pantheon\TerminusBuildTools\API\Bitbucket\BitbucketAPITrait;
 
 use GuzzleHttp\Client;
 
@@ -22,17 +22,14 @@ class BitbucketProvider implements GitProvider, LoggerAwareInterface, Credential
 {
     use LoggerAwareTrait;
     use ExecWithRedactionTrait;
+    use BitbucketAPITrait;
 
     const SERVICE_NAME = 'bitbucket';
     const BITBUCKET_URL = 'https://bitbucket.org';
-    const BITBUCKET_USER = 'BITBUCKET_USER';
-    const BITBUCKET_PASS = 'BITBUCKET_PASS';
     const BITBUCKET_AUTH = 'BITBUCKET_AUTH';
 
     private $bitbucketClient;
     protected $repositoryEnvironment;
-    protected $bitBucketUser;
-    protected $bitBucketPassword;
 
     public function __construct()
     {
@@ -52,94 +49,6 @@ class BitbucketProvider implements GitProvider, LoggerAwareInterface, Credential
         return $this->repositoryEnvironment;
     }
 
-    public function hasToken()
-    {
-        $repositoryEnvironment = $this->getEnvironment();
-        return $repositoryEnvironment->hasToken();
-    }
-
-    public function token()
-    {
-        $repositoryEnvironment = $this->getEnvironment();
-        return $repositoryEnvironment->token();
-    }
-
-    public function setToken($token)
-    {
-        $repositoryEnvironment = $this->getEnvironment();
-        $repositoryEnvironment->setToken(self::BITBUCKET_AUTH, $token);
-    }
-
-    public function getBitBucketUser()
-    {
-        return $this->bitBucketUser;
-    }
-
-    public function getBitBucketPassword()
-    {
-        return $this->bitBucketPassword;
-    }
-
-    public function setBitBucketUser($u)
-    {
-        $this->bitBucketUser = $u;
-    }
-
-    public function setBitBucketPassword($pw)
-    {
-        $this->bitBucketPassword = $pw;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function credentialRequests()
-    {
-        // Tell the credential manager that we require two credentials
-        $bitbucketUserRequest = new CredentialRequest(
-            self::BITBUCKET_USER,
-            "",
-            "Enter your Bitbucket username",
-            '#^.+$#',
-            ""
-        );
-
-        $bitbucketPassRequest = new CredentialRequest(
-            self::BITBUCKET_PASS,
-            "",
-            "Enter your Bitbucket account password or an app password",
-            '#^.+$#',
-            ""
-        );
-
-        return [ $bitbucketUserRequest, $bitbucketPassRequest ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function setCredentials(CredentialProviderInterface $credentials_provider)
-    {
-        // Since the `credentialRequests()` method declared that we need a
-        // BITBUCKET_USER and BITBUCKET_PASS credentials, it will be available
-        // for us to copy from the credentials provider when this method is called.
-        $this->setBitBucketUser($credentials_provider->fetch(self::BITBUCKET_USER));
-        $this->setBitBucketPassword($credentials_provider->fetch(self::BITBUCKET_PASS));
-        $this->setToken(
-            urlencode($this->getBitBucketUser())
-            .':'.
-            urlencode($this->getBitBucketPassword())
-        );
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function authenticatedUser()
-    {
-        return $this->getBitBucketUser();
-    }
-
     /**
      * @inheritdoc
      */
@@ -157,7 +66,7 @@ class BitbucketProvider implements GitProvider, LoggerAwareInterface, Credential
 
         // Create a Bitbucket repository
         $this->logger->notice('Creating repository {repo}', ['repo' => $target_project]);
-        $result = $this->bitbucketAPI('repositories/'.$target_project, 'PUT');
+        $result = $this->api()->request("repositories/$target_project", [], 'PUT');
 
         // Create a git repository. Add an origin just to have the data there
         // when collecting the build metadata later. We use the 'pantheon'
@@ -202,18 +111,19 @@ class BitbucketProvider implements GitProvider, LoggerAwareInterface, Credential
      */
     public function commentOnCommit($target_project, $commit_hash, $message)
     {
-        // We're using the 1.0 API here because this is not yet supported in 2.0
-        $url = "/1.0/repositories/$target_project/changesets/$commit_hash/comments";
-        $data = [ 'content' => $message ];
-        $this->bitbucketAPIClient()->post($url, [
-           'form_params' => $data
-        ]);
+        $body = [
+            'content' => [
+                'raw' => $message,
+            ],
+            // 'parent' => '???', // not sure if this is needed
+        ];
+        $result = $this->api()->request("repositories/$target_project/commit/$commit_hash/comments", $body);
     }
 
     /**
      * @inheritdoc
      */
-    function branchesForPullRequests($target_project, $state)
+    function branchesForPullRequests($target_project, $state, $callback = null)
     {
         $stateParameters = [
             'open' => ['OPEN'],
@@ -222,9 +132,9 @@ class BitbucketProvider implements GitProvider, LoggerAwareInterface, Credential
         ];
         if (!isset($stateParameters[$state]))
             throw new TerminusException("branchesForPullRequests - state must be one of: open, closed, all");
-        
-        $data = $this->bitbucketAPI("repositories/$target_project/pullrequests?state="
-            .implode('&state=', $stateParameters[$state]));
+
+        $data = $this->api()->pagedRequest("repositories/$target_project/pullrequests?state="
+            .implode('&state=', $stateParameters[$state]), $callback);
 
         $branchList = array_column(array_map(
             function ($item) {
@@ -238,49 +148,10 @@ class BitbucketProvider implements GitProvider, LoggerAwareInterface, Credential
         return $branchList;
     }
 
-    private function bitbucketAPIClient()
+    public function convertPRInfo($data)
     {
-        if (!isset($this->bitbucketClient))
-            $this->bitbucketClient = new Client([
-                'base_uri' => 'https://api.bitbucket.org/2.0/',
-                'auth' => [ $this->getBitBucketUser(), $this->getBitBucketPassword() ],
-                'headers' => [
-                    'User-Agent' => ProviderEnvironment::USER_AGENT,
-                ]
-            ]);
-        return $this->bitbucketClient;
-    }
-
-    protected function bitbucketAPI($uri, $method = 'GET', $data = [])
-    {
-        $guzzleParams = [];
-        if (!empty($data)) {
-            $guzzleParams['json'] = $data;
-        }
-
-        $this->logger->notice('Call Bitbucket API: {method} {uri}', ['method' => $method, 'uri' => $uri]);
-        
-        $res = $this->bitbucketAPIClient()->request($method, $uri, $guzzleParams);
-        $resultData = json_decode($res->getBody(), true);
-        $httpCode = $res->getStatusCode();
-
-        $errors = [];
-        if (isset($resultData['errors'])) {
-            foreach ($resultData['errors'] as $error) {
-                $errors[] = $error['message'];
-            }
-        }
-        if ($httpCode && ($httpCode >= 300)) {
-            $errors[] = "Http status code: $httpCode";
-        }
-
-        $message = isset($resultData['message']) ? "{$resultData['message']}." : '';
-
-        if (!empty($message) || !empty($errors)) {
-            throw new TerminusException('{service} error: {message} {errors}', ['service' => $service, 'message' => $message, 'errors' => implode("\n", $errors)]);
-        }
-
-        return $resultData;
+        $isClosed = ($data['state'] != 'OPEN');
+        return new PullRequestInfo($data['id'], $isClosed, $data['source']['branch']['name']);
     }
 
     protected function execGit($dir, $cmd, $replacements = [], $redacted = [])
