+++ src/ServiceProviders/RepositoryProviders/RepositoryEnvironment.php
@@ -6,37 +6,7 @@ use Pantheon\TerminusBuildTools\ServiceProviders\ProviderEnvironment;
 
 class RepositoryEnvironment extends ProviderEnvironment
 {
-    protected $token_key = 'TOKEN';
-    protected $serviceName;
     protected $projectId;
-
-    public function hasToken()
-    {
-        return isset($this[$this->token_key]);
-    }
-
-    public function token()
-    {
-        return $this[$this->token_key];
-    }
-
-    public function setToken($key, $token)
-    {
-        $this->token_key = $key;
-        $this[$key] = $token;
-        return $this;
-    }
-
-    public function serviceName()
-    {
-        return $this->serviceName;
-    }
-
-    public function setServiceName($serviceName)
-    {
-        $this->serviceName = $serviceName;
-        return $this;
-    }
 
     /**
      * Project identifier (org/projectname)
@@ -56,8 +26,11 @@ class RepositoryEnvironment extends ProviderEnvironment
         return $this;
     }
 
+    /**
+     * @inheritdoc
+     */
     public function ciState()
     {
-        return $this->getElements([$this->token_key]);
+        return $this->getElements($this->tokens);
     }
 }
